+++ src/main/java/br/ufjf/luidgisarto/trb2/repositories/AvaliadorRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AvaliadorRepository extends JpaRepository<Avaliador, Long> {
+public interface AvaliadorRepository extends JpaRepository<Avaliador, Integer> {
 
 }
