+++ src/main/java/br/ufjf/luidgisarto/trb2/repositories/RevisaoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RevisaoRepository extends JpaRepository<Revisao, Long> {
+public interface RevisaoRepository extends JpaRepository<Revisao, Integer> {
 
 }
